--- a/go.mod
+++ b/go.mod
@@ -1,9 +1,9 @@
 module lukechampine.com/adiantum
 
-go 1.12
+go 1.20
 
 require (
 	github.com/aead/chacha20 v0.0.0-20180709150244-8b13a72661da
-	golang.org/x/crypto v0.0.0-20190530122614-20be4c3c3ed5
-	golang.org/x/sys v0.0.0-20190602015325-4c4f7f33c9ed
+	golang.org/x/crypto v0.23.0
+	golang.org/x/sys v0.20.0
 )