--- a/go.mod
+++ b/go.mod
@@ -3,7 +3,6 @@ module lukechampine.com/adiantum
 go 1.20
 
 require (
-	github.com/aead/chacha20 v0.0.0-20180709150244-8b13a72661da
 	golang.org/x/crypto v0.23.0
 	golang.org/x/sys v0.20.0
 )