--- a/internal/chacha/chacha_386.go
+++ b/internal/chacha/chacha_386.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a license that can be
 // found in the LICENSE file.
 
-// +build 386,!gccgo,!appengine,!nacl
+//go:build gc
 
 package chacha
 
@@ -29,14 +29,17 @@ func initialize(state *[64]byte, key []byte, nonce *[16]byte) {
 }
 
 // This function is implemented in chacha_386.s
+//
 //go:noescape
 func hChaCha20SSE2(out *[32]byte, nonce *[16]byte, key *[32]byte)
 
 // This function is implemented in chacha_386.s
+//
 //go:noescape
 func hChaCha20SSSE3(out *[32]byte, nonce *[16]byte, key *[32]byte)
 
 // This function is implemented in chacha_386.s
+//
 //go:noescape
 func xorKeyStreamSSE2(dst, src []byte, block, state *[64]byte, rounds int) int