--- a/internal/chacha/chacha_amd64.go
+++ b/internal/chacha/chacha_amd64.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a license that can be
 // found in the LICENSE file.
 
-// +build go1.7,amd64,!gccgo,!appengine,!nacl
+//go:build gc
 
 package chacha
 
@@ -16,34 +16,42 @@ func init() {
 }
 
 // This function is implemented in chacha_amd64.s
+//
 //go:noescape
 func initialize(state *[64]byte, key []byte, nonce *[16]byte)
 
 // This function is implemented in chacha_amd64.s
+//
 //go:noescape
 func hChaCha20SSE2(out *[32]byte, nonce *[16]byte, key *[32]byte)
 
 // This function is implemented in chacha_amd64.s
+//
 //go:noescape
 func hChaCha20SSSE3(out *[32]byte, nonce *[16]byte, key *[32]byte)
 
 // This function is implemented in chachaAVX2_amd64.s
+//
 //go:noescape
 func hChaCha20AVX(out *[32]byte, nonce *[16]byte, key *[32]byte)
 
 // This function is implemented in chacha_amd64.s
+//
 //go:noescape
 func xorKeyStreamSSE2(dst, src []byte, block, state *[64]byte, rounds int) int
 
 // This function is implemented in chacha_amd64.s
+//
 //go:noescape
 func xorKeyStreamSSSE3(dst, src []byte, block, state *[64]byte, rounds int) int
 
 // This function is implemented in chacha_amd64.s
+//
 //go:noescape
 func xorKeyStreamAVX(dst, src []byte, block, state *[64]byte, rounds int) int
 
 // This function is implemented in chachaAVX2_amd64.s
+//
 //go:noescape
 func xorKeyStreamAVX2(dst, src []byte, block, state *[64]byte, rounds int) int