--- a/internal/chacha/chacha_amd64.s
+++ b/internal/chacha/chacha_amd64.s
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a license that can be
 // found in the LICENSE file.
 
-// +build amd64,!gccgo,!appengine,!nacl
+//go:build gc
 
 #include "const.s"
 #include "macro.s"