--- a/internal/chacha/chacha_arm64.go
+++ b/internal/chacha/chacha_arm64.go
@@ -1,0 +1,54 @@
+//go:build gc
+
+package chacha
+
+import "encoding/binary"
+
+func init() {
+	useVX = true
+}
+
+func initialize(state *[64]byte, key []byte, nonce *[16]byte) {
+	binary.LittleEndian.PutUint32(state[0:], sigma[0])
+	binary.LittleEndian.PutUint32(state[4:], sigma[1])
+	binary.LittleEndian.PutUint32(state[8:], sigma[2])
+	binary.LittleEndian.PutUint32(state[12:], sigma[3])
+	copy(state[16:], key[:])
+	copy(state[48:], nonce[:])
+}
+
+func xorKeyStream(dst, src []byte, block, state *[64]byte, rounds int) int {
+	switch {
+	case useVX:
+		const bufSize = 256
+
+		key := (*[32]byte)(state[16:48])
+		counter := (*[4]byte)(state[48:52])
+		nonce := (*[12]byte)(state[52:64])
+
+		// Handle N full bufSize blocks.
+		if full := len(src) &^ (bufSize - 1); full > 0 {
+			xorKeyStreamVX(dst[:full], src[:full], key, nonce, counter, rounds)
+			dst, src = dst[full:], src[full:]
+		}
+
+		// Handle a partial bufSize block.
+		if len(src) > 0 {
+			var buf [bufSize]byte
+			copy(buf[:], src)
+			xorKeyStreamVX(buf[:], buf[:], key, nonce, counter, rounds)
+			copy(dst, buf[:])
+		}
+
+		return len(src)
+	default:
+		return xorKeyStreamGeneric(dst, src, block, state, rounds)
+	}
+}
+
+func hChaCha20(out *[32]byte, nonce *[16]byte, key *[32]byte) {
+	hChaCha20Generic(out, nonce, key)
+}
+
+//go:noescape
+func xorKeyStreamVX(dst, src []byte, key *[32]byte, nonce *[12]byte, counter *[4]byte, rounds int)