--- a/internal/chacha/chacha_arm64.go
+++ b/internal/chacha/chacha_arm64.go
@@ -2,19 +2,12 @@
 
 package chacha
 
-import "encoding/binary"
-
 func init() {
 	useVX = true
 }
 
 func initialize(state *[64]byte, key []byte, nonce *[16]byte) {
-	binary.LittleEndian.PutUint32(state[0:], sigma[0])
-	binary.LittleEndian.PutUint32(state[4:], sigma[1])
-	binary.LittleEndian.PutUint32(state[8:], sigma[2])
-	binary.LittleEndian.PutUint32(state[12:], sigma[3])
-	copy(state[16:], key[:])
-	copy(state[48:], nonce[:])
+	initializeGeneric(state, key, nonce)
 }
 
 func xorKeyStream(dst, src []byte, block, state *[64]byte, rounds int) int {