--- a/internal/chacha/chacha_ref.go
+++ b/internal/chacha/chacha_ref.go
@@ -2,18 +2,11 @@
 // Use of this source code is governed by a license that can be
 // found in the LICENSE file.
 
-//go:build !(gc && (386 || amd64))
+//go:build !(gc && (386 || amd64 || arm64))
 
 package chacha
 
 import "encoding/binary"
-
-func init() {
-	useSSE2 = false
-	useSSSE3 = false
-	useAVX = false
-	useAVX2 = false
-}
 
 func initialize(state *[64]byte, key []byte, nonce *[16]byte) {
 	binary.LittleEndian.PutUint32(state[0:], sigma[0])