… adiantum_test.go
… go.sum
… hpolyc/hpolyc_test.go
… internal/chacha/chacha_generic.go
… internal/chacha/chacha_test.go
… internal/chacha/const.s
… internal/chacha/macro.s
… internal/xchacha/chacha.go
… nh/nhSSE2_amd64.s
… nh/nh_amd64.go
… nh/nh_amd64.s
… nh/nh_generic.go
… nh/nh_little.go
… nh/nh_test.go