Hi! 👋

This branch includes a slew of updates that significantly improve performance on [`arm64`](https://github.com/ncruces/adiantum/actions/runs/9009610320/job/24754159188).

Before:
```
BenchmarkAdiantum/XChaCha12_Encrypt-3        	  113114	     11517 ns/op	 355.66 MB/s	       0 B/op	       0 allocs/op
BenchmarkAdiantum/XChaCha12_Decrypt-3        	  105201	     11246 ns/op	 364.22 MB/s	       0 B/op	       0 allocs/op
```

After:
```
BenchmarkAdiantum/XChaCha12_Encrypt-3        	  285481	      4209 ns/op	 973.20 MB/s	       0 B/op	       0 allocs/op
BenchmarkAdiantum/XChaCha12_Decrypt-3        	  285500	      4224 ns/op	 969.65 MB/s	       0 B/op	       0 allocs/op
```

You may want to consider each commit individually, or reject this altogether. I really don't mind doing some additional work to upstream this! This is just where I'm at, and I understand some of it is opinionated.

Commits:
1. 5028739f27d5118ca4212b2f13b86f15840f473f updates dependencies, and set Go to 1.20 (needed by c8607862f781328052ea21cb93d4656cf6ae6dce)
2. 852cde011c6466d058acf2f5a2ad27a000301876 `github.com/aead/chacha20` is [archived](https://github.com/aead/chacha20/), vendor it
3. 878546e283872a5f29eea11cc7690f6ea2e521d0 removes the bits that aren't used from 852cde011c6466d058acf2f5a2ad27a000301876
4. 8ae39ea1182fe34430331278836d46eb8549464c updates to use `go:build` rather than `+build`, and revises tags (`nacl` doesn't exist anymore, etc)
5. 640c72074dea998ad2a166a365d73bb7e721a7a3 adapts `arm64` assembly from `golang.org/x/crypto/chacha20`
6. e3261e2e06f26ae14c0fe981d692533cd6cc8900 refactor `chacha` for code reuse
7. b6059c569cae0e49e810023b54cd803d77367d40 improves the fallback version of `nh.Sum`
8. c8607862f781328052ea21cb93d4656cf6ae6dce adds little endian version of `nh.Sum`
9. cb1b8cc2c8e186be76d79373f1a262a54dbe2dd5 avoids deprecated `ioutil`

Commit 640c72074dea998ad2a166a365d73bb7e721a7a3 is a bit hacked up, and might benefit from additional work. I just copied the assembly, and made the rounds configurable. Didn't even bother with fixing the setup parts. OTOH, this means that if upstream has fixes, they're easier to integrate. Also, the same interface is used by `ppc64` and `s390x` assembly.

Commit c8607862f781328052ea21cb93d4656cf6ae6dce is a bit of a cop out. Instead of sitting down and writing `arm64` assembly, I just coded a Go version that should work on any little endian architecture, and used `unsafe` for setup. Still, it's about as fast `sumAsm` on `amd64`, and produces decent code on a wide range of architectures. This is responsible for much of the speed up on Adiantum.

Commit b6059c569cae0e49e810023b54cd803d77367d40 improves the situation a bit for the big endian architectures left out in c8607862f781328052ea21cb93d4656cf6ae6dce, by unrolling a loop and avoiding some bounds checks.