@lukechampine 2024-05-10
Thanks again for this. I hope Go (and [avo](https://github.com/mmcloughlin/avo)) get better support for ARM assembly soon; I would very much like to accelerate `blake3` in the same way.